fix(security): disable external data transmission to ntfy.sh
- Replace notification service with security-patched version
- Block all external notification transmission
- Force localhost-only URLs for notifications
- Add comprehensive security analysis documentation
- Prevent metadata leakage of Claude CLI activity

BREAKING CHANGE: External notifications to ntfy.sh are now completely disabled.
Users must self-host ntfy or use alternative local notification methods.
